+++ src/Sulu/Bundle/ContentBundle/Tests/Functional/Controller/NodeControllerTest.php
@@ -291,14 +291,17 @@ class NodeControllerTest extends SuluTestCase
         $targetPage = $this->documentManager->create('page');
         $targetPage->setTitle('target');
         $targetPage->setResourceSegment('/target');
+        $targetPage->setStructureType('default');
         $this->documentManager->persist($targetPage, 'en', array('parent_path' => '/cmf/sulu_io/contents'));
 
         $this->documentManager->flush();
 
         $internalLinkPage = $this->documentManager->create('page');
         $internalLinkPage->setTitle('page');
+        $internalLinkPage->setStructureType('internal-link');
         $internalLinkPage->setRedirectType(RedirectType::INTERNAL);
         $internalLinkPage->setRedirectTarget($targetPage);
+        $internalLinkPage->setResourceSegment('/test');
         $this->documentManager->persist($internalLinkPage, 'en', array('parent_path' => '/cmf/sulu_io/contents'));
 
         $this->documentManager->flush();
@@ -316,8 +319,10 @@ class NodeControllerTest extends SuluTestCase
 
         $externalLinkPage = $this->documentManager->create('page');
         $externalLinkPage->setTitle('page');
+        $externalLinkPage->setStructureType('external-link');
         $externalLinkPage->setRedirectType(RedirectType::EXTERNAL);
         $externalLinkPage->setRedirectExternal('http://www.sulu.io');
+        $externalLinkPage->setResourceSegment('/test');
         $this->documentManager->persist($externalLinkPage, 'en', array('parent_path' => '/cmf/sulu_io/contents'));
 
         $this->documentManager->flush();