+++ src/Sulu/Component/Webspace/Loader/XmlFileLoader.php
@@ -12,7 +12,9 @@ namespace Sulu\Component\Webspace\Loader;
 
 use Sulu\Component\Localization\Localization;
 use Sulu\Component\Webspace\Environment;
+use Sulu\Component\Webspace\Loader\Exception\ExpectedDefaultTemplatesNotFound;
 use Sulu\Component\Webspace\Loader\Exception\InvalidAmountOfDefaultErrorTemplateException;
+use Sulu\Component\Webspace\Loader\Exception\InvalidAmountOfDefaultTemplatesException;
 use Sulu\Component\Webspace\Loader\Exception\InvalidDefaultErrorTemplateException;
 use Sulu\Component\Webspace\Loader\Exception\InvalidDefaultLocalizationException;
 use Sulu\Component\Webspace\Loader\Exception\InvalidErrorTemplateException;
@@ -307,6 +309,7 @@ class XmlFileLoader extends FileLoader
         }
 
         $this->generateErrorTemplates($theme);
+        $this->generateDefaultTemplates($theme);
 
         return $theme;
     }
@@ -337,6 +340,30 @@ class XmlFileLoader extends FileLoader
         // only one or none default error-template is legal
         if ($defaultErrorTemplates > 1) {
             throw new InvalidAmountOfDefaultErrorTemplateException($this->webspace->getKey());
+        }
+
+        return $theme;
+    }
+
+    private function generateDefaultTemplates(Theme $theme)
+    {
+        $expected = array('page', 'homepage');
+        $found = array();
+        $nodes = $this->xpath->query('/x:webspace/x:theme/x:default-templates/x:default-template');
+
+        foreach ($nodes as $node) {
+            /** @var \DOMNode $node */
+            $template = $node->nodeValue;
+            $type = $node->attributes->getNamedItem('type')->nodeValue;
+
+            $theme->addDefaultTemplate($type, $template);
+            $found[] = $type;
+        }
+
+        foreach ($expected as $item) {
+            if (!in_array($item, $found)) {
+                throw new ExpectedDefaultTemplatesNotFound($this->webspace->getKey(), $expected, $found);
+            }
         }
 
         return $theme;