+++ CHANGELOG.md
@@ -1,13 +1,18 @@
 CHANGELOG for Sulu
 ==================
+
 * dev-develop
+    * FEATURE     #1273 [SnippetBundle]  Removed snippet state from ui and set default published
     * BUGFIX      #1308 [ContactBundle]  Fixed adding a new contact-account relation in contacts tab of account
     * ENHANCEMENT #1302 [MediaBundle]    Removed 'partial-load' hint from getMediaById repository function.
-    * BUGFIX      #1172 [ContentBundle]  Fix hreflang meta tags, remove unvalid title meta tag
     * BUGFIX      #1172 [ContentBundle]  Fix hreflang meta tags, remove invalid title meta tag
     * BUGFIX      #1172 [ContentBundle]  Fixed ckeditor in blocks for IE11
     * ENHANCEMENT #1040 [ContentBundle]  Added validation for required tag name
     * FEATURE     #1273 [SnippetBundle]  Removed snippet state from ui and set default published
+    * FEATURE     #1281 [ContentBundle]  Added default-template config to webspace theme
+
+* dev-master
+    * HOTFIX      #1306 [ContentBundle]  Fixed migration commands for jackrabbit
 
 * 1.0.0-RC3 (2015-06-24)
     * HOTFIX      #1306 [ContentBundle]  Fixed migration commands for jackrabbit